Skips tests that won't work on Safari Mobile due to the iframe added by Karma
Safari Mobile has a weird behavior for iframes that is different from all other browsers. This doesn't allow us to test scrolling the document on Safari Mobile, since Karma runs the test inside an iframe. We've tried to tell Karma not to use the iframe, but in this case it requires opening pop ups, which are disabled on mobile phone browsers in saucelabs, so that doesn't work either. Unfortunately, the only way is to skip the tests related to document scrolling on Safari Mobile for now.