Manually build data for the first render when not ready yet
When we render for the first time we don't have any data, since senna hasn't navigated anywhere yet. We can't wait for senna to properly navigate as that's async, and rendering the router has to be sync on the first time if the user requests so. This is just manually calling "data" on the render method when necessary for now.