Fixes call to "skipNextChildrenDisposal"
The api for this has changed, it needs to receive the
component instance. There's also some internal data that was being accessed which has changed. Accessing these is dangerous, we need to find a way to provide this feature without these hacks. For now this is just fixing the router though, to unblock its usage.